fixed more issues with fees

+++ merk/src/merk/mod.rs
@@ -27,6 +27,7 @@ use crate::{
         NULL_HASH,
     },
 };
+use crate::MerkType::{BaseMerk, StandaloneMerk, LayeredMerk};
 
 type Proof = (LinkedList<ProofOp>, Option<u16>, Option<u16>);
 
@@ -152,11 +153,32 @@ impl<'a, I: RawIterator> KVIterator<'a, I> {
     }
 }
 
+#[derive(PartialEq, Eq)]
+pub enum MerkType {
+    /// A StandaloneMerk has it's root key storage on a field and pays for root key updates
+    StandaloneMerk,
+    /// A BaseMerk has it's root key storage on a field but does not pay for when these keys change
+    BaseMerk,
+    /// A LayeredMerk has it's root key storage inside a parent merk
+    LayeredMerk,
+}
+
+impl MerkType {
+    pub(crate) fn requires_root_storage_update(&self) -> bool {
+        match self {
+            StandaloneMerk => true,
+            BaseMerk => true,
+            LayeredMerk => false
+        }
+    }
+}
+
 /// A handle to a Merkle key/value store backed by RocksDB.
 pub struct Merk<S> {
     pub(crate) tree: Cell<Option<Tree>>,
     pub(crate) root_tree_key: Cell<Option<Vec<u8>>>,
     pub storage: S,
+    pub merk_type: MerkType,
 }
 
 impl<S> fmt::Debug for Merk<S> {
@@ -181,12 +203,24 @@ where
     S: StorageContext<'db>,
     <S as StorageContext<'db>>::Error: std::error::Error,
 {
-    pub fn open_empty(storage: S) -> Self {
+    pub fn open_empty(storage: S, merk_type: MerkType) -> Self {
         Self {
             tree: Cell::new(None),
             root_tree_key: Cell::new(None),
             storage,
+            merk_type,
         }
+    }
+
+    pub fn open_standalone(storage: S) -> CostContext<Result<Self>> {
+        let mut merk = Self {
+            tree: Cell::new(None),
+            root_tree_key: Cell::new(None),
+            storage,
+            merk_type: StandaloneMerk
+        };
+
+        merk.load_base_root().map_ok(|_| merk)
     }
 
     pub fn open_base(storage: S) -> CostContext<Result<Self>> {
@@ -194,16 +228,18 @@ where
             tree: Cell::new(None),
             root_tree_key: Cell::new(None),
             storage,
+            merk_type: BaseMerk
         };
 
         merk.load_base_root().map_ok(|_| merk)
     }
 
-    pub fn open_with_root_key(storage: S, root_key: Option<Vec<u8>>) -> CostContext<Result<Self>> {
+    pub fn open_layered_with_root_key(storage: S, root_key: Option<Vec<u8>>) -> CostContext<Result<Self>> {
         let mut merk = Self {
             tree: Cell::new(None),
             root_tree_key: Cell::new(root_key),
             storage,
+            merk_type: MerkType::LayeredMerk
         };
 
         merk.load_root().map_ok(|_| merk)
@@ -541,12 +577,15 @@ where
         }
 
         Walker::apply_to(maybe_walker, batch, self.source()).flat_map_ok(
-            |(maybe_tree, updated_keys, deleted_keys)| {
+            |(maybe_tree, new_keys, updated_keys, deleted_keys, updated_root_key_from)| {
+                // we set the new root node of the merk tree
                 self.tree.set(maybe_tree);
                 // commit changes to db
                 self.commit(
+                    new_keys,
                     updated_keys,
                     deleted_keys,
+                    updated_root_key_from,
                     aux,
                     update_tree_value_based_on_costs,
                     section_removal_bytes,
@@ -643,8 +682,10 @@ where
 
     pub fn commit<K>(
         &mut self,
+        new_keys: BTreeSet<Vec<u8>>,
         updated_keys: BTreeSet<Vec<u8>>,
         deleted_keys: LinkedList<Vec<u8>>,
+        updated_root_key_from: Option<Vec<u8>>,
         aux: &AuxMerkBatch<K>,
         update_tree_value_based_on_costs: &mut impl FnMut(
             &StorageCost,
@@ -682,26 +723,41 @@ where
                     None => {}
                     Some(_) => {}
                 }
-                let costs = if updated_keys.contains(tree_key) {
-                    Some(KeyValueStorageCost::for_updated_root_cost(
-                        tree_key.len() as u32
-                    ))
-                } else {
-                    None
-                };
-                // update pointer to root node
-                cost_return_on_error_no_add!(
-                    &inner_cost,
-                    batch
-                        .put_root(ROOT_KEY_KEY, tree_key, costs)
-                        .map_err(CostError)
-                        .map_err(|e| e.into())
-                );
+                // if they are a base merk we should update the root key
+                if self.merk_type.requires_root_storage_update() {
+                    // there are two situation where we want to put the root key
+                    // it was updated from something else
+                    // or it is part of new keys
+                    if updated_root_key_from.is_some() || new_keys.contains(tree_key) {
+
+                        let costs = if self.merk_type == StandaloneMerk {
+                            // if we are a standalone merk we want real costs
+                             Some(KeyValueStorageCost::for_updated_root_cost(updated_root_key_from.as_ref().map(|k| k.len() as u32),
+                                                                                   tree_key.len() as u32))
+                        } else {
+                            // if we are a base merk we estimate these costs are free
+                            // This None does not guarantee they are free though
+                            None
+                        };
+
+
+                        // update pointer to root node
+                        cost_return_on_error_no_add!(
+                                    &inner_cost,
+                                    batch
+                                        .put_root(ROOT_KEY_KEY, tree_key, costs)
+                                        .map_err(CostError)
+                                        .map_err(|e| e.into())
+                                );
+                    }
+                }
 
                 Ok(committer.batch)
             } else {
-                // empty tree, delete pointer to root
-                batch.delete_root(ROOT_KEY_KEY);
+                if self.merk_type.requires_root_storage_update() {
+                    // empty tree, delete pointer to root
+                    batch.delete_root(ROOT_KEY_KEY);
+                }
 
                 Ok(vec![])
             }
